WIP: Time Parallelism
This implements time parallelism by adding a few options in config. When running fdtd-acoustic-simulator-viewer, pass in the option --index [index] to have the current run run a specific chunk. I added an example config file. I've verified this on a modal example, but it should work for kinematic or other examples. 

This is unfinished because currently the modal ODE computes the effects of all the impulses. A future change should make it cull impulses that are too early or too late. However, I found that for most examples, it doesn't take too much time for the Modal ODE to step through all the impulses, because the computations are on q rather than x.